Continue development:
    * Fixed bug: File > Reload creates a new tab
    * New dialog for metronome in Menu > Tools, to facilitate more features,
      such as setting tempo by tapping.
    * Added help system. Help menu now includes: users guide, study guide
      and editor quick guide.
    * Added class HelpSystem to isolate implementation details and facilitate
      experimentation with
      different alternatives
    * Added code for importing MusicXML files. Importing files merged in
      File > Open.
    * Designed and tested the workflow for translating Users' Guide and
      Developers' Guide. Several scripts developed for managing in this.
    * Added option, in preferences dialog, for enabling experimental features.
    * Fixed bug 80207 - program freezes when inserting score using Score Wizard
    * Implemented script command 'sel ?'
    * Added code for creating chords. Shortcut 'Shift + [a-g]'
    * Shorcuts:
- Added code for global management of shortcuts
- Added option dialog for customizing shortcuts
    * Toobox Redesigned:
- Added Page Selector control
- Shortcuts for note duration changed to use the same than Finale
          and MuseScore.
- Removed group for selecting note or rest.
    * Added code for clearing current selection. Shortcut 'Esc'
    * Added console commands for managing selections.
    * Fixed bug in PageSelector causing a crash. Menu id was used,
      incorrectly, as page id.
    * Code updated for using wxWidgets 3.0

+++ include/lenmus_tool_page_clefs.h
@@ -1,6 +1,6 @@
 //---------------------------------------------------------------------------------------
 //    LenMus Phonascus: The teacher of music
-//    Copyright (c) 2002-2014 LenMus project
+//    Copyright (c) 2002-2015 LenMus project
 //
 //    This program is free software; you can redistribute it and/or modify it under the
 //    terms of the GNU General Public License as published by the Free Software Foundation,
@@ -180,7 +180,7 @@ public:
 
 protected:
     //mandatory overrides
-    int get_key_translation_context() { return k_key_context_clefs; }
+    long get_key_translation_context() { return k_key_context_clefs; }
 
 
 //    //groups
