+++ include/lenmus_tool_page_symbols.h
@@ -1,6 +1,6 @@
 //---------------------------------------------------------------------------------------
 //    LenMus Phonascus: The teacher of music
-//    Copyright (c) 2002-2014 LenMus project
+//    Copyright (c) 2002-2015 LenMus project
 //
 //    This program is free software; you can redistribute it and/or modify it under the
 //    terms of the GNU General Public License as published by the Free Software Foundation,
@@ -110,7 +110,7 @@ public:
 
 private:
     //mandatory overrides
-    int get_key_translation_context() { return k_key_context_symbols; }
+    long get_key_translation_context() { return k_key_context_symbols; }
 
 //    //groups
 //    GrpSymbols*           m_pGrpSymbols;
